backend/mcp-server/requirements.txt
mcp>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
